ReducedComplexityModeling

Documentation for ReducedComplexityModeling.

ReducedComplexityModeling.AutoEncoderModel

ReducedComplexityModeling.CartesianParameterSampler

ReducedComplexityModeling.Parameter

ReducedComplexityModeling.ParameterSampler

ReducedComplexityModeling.ParameterSpace

ReducedComplexityModeling.ReducedBasisModel

ReducedComplexityModeling.h5load

ReducedComplexityModeling.h5save

ReducedComplexityModeling.learn

ReducedComplexityModeling.read_parameters

ReducedComplexityModeling.sample

ReducedComplexityModeling.save_parameters

ReducedComplexityModeling.AutoEncoderModel — Type

Autoencoders

Uses Variational Autoencoders to represent the data in its reduced state.

ReducedComplexityModeling.CartesianParameterSampler — Type

ReducedComplexityModeling.Parameter — Type

ReducedComplexityModeling.ParameterSampler — Type

ReducedComplexityModeling.ParameterSpace — Type

ParameterSpace collects all parameters of a system as well as samples in the parameter space.

ReducedComplexityModeling.ReducedBasisModel — Type

Reduced Basis

Uses ModelOrderReduction methods from ReducedBasisMethods.jl to represent the data in latent space.

ReducedComplexityModeling.h5load — Method

Load ParameterSpace

ReducedComplexityModeling.h5save — Method

save ParameterSpace

ReducedComplexityModeling.learn — Method

learn(prob)

Uses data from high-fidelity simulations/experiments and learns the latent representation of the problem.

ReducedComplexityModeling.read_parameters — Function

read parameters

ReducedComplexityModeling.sample — Method

ReducedComplexityModeling.save_parameters — Method

save parameters